# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AsamApp API
  description: |
    API REST para gestion de eventos gastronomicos: cotizaciones, inventario,
    compras, finanzas, notas y operaciones. Backend Node.js + Express + Prisma + SQLite.
  version: 1.0.0
  contact:
    name: ProyectoAsado
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Iniciar sesion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login exitoso
        '401':
          description: Credenciales invalidas
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registrar primer usuario (solo si no hay usuarios)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email: { type: string }
                username: { type: string }
                password: { type: string, minLength: 4 }
      responses:
        '201':
          description: Usuario creado como admin
        '403':
          description: Registro deshabilitado
...